[mixin] Moved download `crl-view` to mixin #116
Fixes #116
